+++ src/server/scripts/Commands/cs_cheat.cpp
@@ -232,7 +232,7 @@ public:
                 ChatHandler(chr->GetSession()).PSendSysMessage(LANG_YOURS_EXPLORE_SET_NOTHING, handler->GetNameLink().c_str());
         }
 
-        for (uint8 i = 0; i < PLAYER_EXPLORED_ZONES_SIZE; ++i)
+        for (uint32 i = 0; i < PLAYER_EXPLORED_ZONES_SIZE; ++i)
         {
             if (reveal)
                 handler->GetSession()->GetPlayer()->SetFlag(PLAYER_EXPLORED_ZONES_1+i, 0xFFFFFFFF);
